requirements.txt
numpy==2.1.3
pygame==2.6.1
PyOpenGL==3.1.7
python-dotenv==1.0.1
requests==2.32.3
pynput==1.7.7
python-osc==1.9.0
